policy for employee and banks

+++ app/controllers/api/v2/admin/banks_controller.rb
@@ -1,25 +1,23 @@
 class Api::V2::Admin::BanksController < Api::V2::Admin::BaseController
   def index
-    banks = Bank.all.order(:id)
-    # authorize [:admin, banks]
+    banks = policy_scope [:admin, Bank]
     paginate json: banks.as_json(only: [:id, :code, :name, :slug, :description]), status: 200
   end
 
   def allbanks
-    banks = Bank.all.order(:id)
-    # authorize [:admin, banks]
+    banks = policy_scope [:admin, Bank]
     render json: banks, status: 200    
   end
 
   def show
     bank = Bank.find(params[:id])
-    # authorize [:admin, banks]
+    authorize [:admin, bank]
     render json: bank, status: 200
   end
 
   def create
     bank = Bank.new(bank_params)
-    # authorize [:admin, banks]
+    authorize [:admin, bank]
     if bank.save
       render json: bank, status: 201
     else
@@ -29,7 +27,7 @@ class Api::V2::Admin::BanksController < Api::V2::Admin::BaseController
 
   def update
     bank = Bank.find(params[:id])
-    # authorize [:admin, banks]
+    authorize [:admin, bank]
     if bank.update(bank_params)
       render json: bank, status: 200
     else
@@ -39,7 +37,7 @@ class Api::V2::Admin::BanksController < Api::V2::Admin::BaseController
 
   def destroy
     bank = Bank.find(params[:id])
-    # authorize [:admin, banks]
+    authorize [:admin, bank]
     bank.destroy
     head 204
   end
@@ -47,6 +45,6 @@ class Api::V2::Admin::BanksController < Api::V2::Admin::BaseController
   private
 
   def bank_params
-    params.permit(:code, :name, :slug, :description)
+    params.permit(policy([:admin, Bank]).permitted_attributes)
   end
 end
